fixed the previous clean up
