+++ data/DISTRIBUTED_IMPROVEMENTS.md
@@ -1,8 +1,8 @@
-# 🚀 Distributed Processing Improvements
+# ✅ Codebase Cleanup Complete
 
-## Summary of Changes
+## Summary
 
-Your NASA NEO data processor has been upgraded with **fully distributed processing capabilities** that eliminate ALL single-threaded bottlenecks using Apache Spark.
+Your NASA NEO data processor has been **cleaned up and streamlined** to use only the high-performance distributed version. All single-threaded code has been removed, resulting in a minimal, production-ready codebase.
 
 ## 🎯 Problems Solved
 
@@ -17,38 +17,37 @@ Your NASA NEO data processor has been upgraded with **fully distributed processi
 3. **Distributed Joins**: Replace Python loops with Spark distributed joins
 4. **Comprehensive Aggregations**: Enhanced statistics calculated in parallel
 
-## 📁 New Files Created
+## 📁 Final Clean Structure
 
 ```
-src/distributed_api_client.py       # Parallel API calls using Spark
-src/distributed_data_processor.py   # Fully distributed data processing
-src/distributed_neo_processor.py    # Distributed pipeline orchestrator
-neo_data_processor_distributed.py   # New entry point with benchmarking
+src/
+├── api_client.py      # Distributed NASA API client using Spark
+├── data_processor.py  # Distributed data processing using Spark  
+├── neo_processor.py   # Distributed pipeline orchestrator
+├── config.py          # Configuration management
+├── models.py          # Data models and exceptions
+├── storage.py         # Data lake operations
+└── __init__.py        # Package initialization
+neo_data_processor.py   # Main entry point (1,905 total lines of code)
 ```
 
 ## 🚀 How to Use
 
-### 1. Run Distributed Processing
+### 1. Run Processing Pipeline
 ```bash
 source venv/bin/activate
-./venv/bin/python neo_data_processor_distributed.py
+./venv/bin/python neo_data_processor.py
 ```
 
-### 2. Performance Comparison
-```bash
-# Compare single-threaded vs distributed
-./venv/bin/python neo_data_processor_distributed.py --compare
-```
-
-### 3. Benchmark Different Sizes
+### 2. Performance Benchmarking
 ```bash
 # Run comprehensive benchmark
-./venv/bin/python neo_data_processor_distributed.py --benchmark
+./venv/bin/python neo_data_processor.py --benchmark
 ```
 
-### 4. Custom Configuration
+### 3. Custom Configuration
 ```python
-from src.distributed_neo_processor import process_neo_data_distributed
+from src.neo_processor import process_neo_data_distributed
 
 result = process_neo_data_distributed(
     limit=200,          # Number of objects to process
